refactor: centralizar envio de WhatsApp
