+++ app/routes.py
@@ -4,11 +4,12 @@ import os
 import json
 import logging
 import uuid
+import random
+import time
 import requests
 from urllib.parse import urljoin
 from app.processamento.mapear_gerencia import mapear_equipe
 from app.processamento.csv_reader import carregar_dados
-from app.tasks import enqueue_csv_processing, get_task_status
 from app.config.settings import (
     EVOLUTION_INSTANCE,
     EVOLUTION_TOKEN,
@@ -24,6 +25,107 @@ def _evo_headers():
     return {"Content-Type": "application/json", "apikey": EVOLUTION_TOKEN}
 
 
+def verificar_sessao() -> bool:
+    """Garante que a sessão do WhatsApp esteja ativa."""
+    url = urljoin(
+        EVOLUTION_URL,
+        f"/instance/fetchInstances?instanceName={EVOLUTION_INSTANCE}",
+    )
+    try:
+        resp = requests.get(url, headers=_evo_headers(), timeout=30)
+        resp.raise_for_status()
+        data = resp.json()
+        estado = (
+            data.get("state")
+            or data.get("connectionState")
+            or data.get("instance", {}).get("state")
+        )
+        if isinstance(estado, str) and estado.lower() == "open":
+            return True
+    except Exception as exc:  # noqa: BLE001
+        logging.error("Erro ao verificar instâncias: %s", exc)
+
+    url = urljoin(EVOLUTION_URL, f"/instance/connect/{EVOLUTION_INSTANCE}")
+    try:
+        resp = requests.get(url, headers=_evo_headers(), timeout=30)
+        resp.raise_for_status()
+        data = resp.json()
+        estado = (
+            data.get("state")
+            or data.get("connectionState")
+            or data.get("instance", {}).get("state")
+        )
+        return isinstance(estado, str) and estado.lower() == "open"
+    except Exception as exc:  # noqa: BLE001
+        logging.error("Erro ao conectar instância: %s", exc)
+        return False
+
+
+def enviar_whatsapp(numero, mensagem, equipe=None):
+    if not verificar_sessao():
+        logging.error("Sessão do WhatsApp desconectada")
+        raise RuntimeError("Sessão do WhatsApp desconectada")
+
+    numero_formatado = numero.replace("+", "").replace("-", "").replace(" ", "")
+    url = urljoin(EVOLUTION_URL, f"/message/sendText/{EVOLUTION_INSTANCE}")
+
+    payload = {
+        "number": numero_formatado,
+        "text": mensagem,
+        "delay": 250,
+    }
+
+    try:
+        logging.info("⏳ Enviando para %s (Equipe: %s)", numero_formatado, equipe)
+        logging.info("Payload: %s", payload)
+
+        response = requests.post(
+            url, json=payload, headers=_evo_headers(), timeout=30
+        )
+
+        logging.info("Evolution API status: %s", response.status_code)
+        logging.info("Evolution API response: %s", response.text)
+
+        if 400 <= response.status_code < 500:
+            req = response.request
+            logging.error(
+                "Falha 4xx ao chamar Evolution API - endpoint=%s método=%s body=%s",
+                req.url,
+                req.method,
+                req.body,
+            )
+
+        if response.status_code not in [200, 201]:
+            raise Exception(
+                f"Erro Evolution API: {response.status_code} - {response.text}"
+            )
+
+        response_data = response.json()
+        if not response_data.get("success", True):
+            raise Exception(
+                f"Erro na resposta: {response_data.get('message', 'Erro desconhecido')}"
+            )
+
+        logging.info(
+            "✅ Mensagem enviada para %s (Equipe: %s)", numero_formatado, equipe
+        )
+
+        time.sleep(random.uniform(0.25, 0.5))
+
+    except SystemExit as se:
+        logging.error(
+            "🚨 SYSTEMEXIT CAPTURADO - Worker sendo morto pelo Gunicorn!"
+        )
+        logging.error("🚨 Exit code: %s", se.code)
+        raise
+    except Exception as e:  # noqa: BLE001
+        logging.error("❌ Falha ao enviar para %s - %s", numero_formatado, e)
+        raise
+    except BaseException as be:  # noqa: BLE001
+        logging.error("🚨 BASEEXCEPTION CAPTURADA: %s", type(be).__name__)
+        raise
+
+
 @api_bp.route('/config', methods=['GET'])
 def get_config():
     return jsonify({
@@ -33,6 +135,7 @@ def get_config():
 
 @api_bp.route('/enviar', methods=['POST'])
 def enviar():
+    from app.tasks import enqueue_csv_processing
     try:
         file = request.files.get('csvFile')
         ignorar_sabados = request.form.get('ignorarSabados', 'true') == 'true'
@@ -80,6 +183,8 @@ def enviar():
 @api_bp.route('/status/<task_id>', methods=['GET'])
 def status(task_id):
     """Retorna o andamento e o resultado de uma tarefa agendada."""
+    from app.tasks import get_task_status
+
     task = get_task_status(task_id)
     if not task:
         return jsonify({"success": False, "error": "Tarefa não encontrada"}), 404
